// frontend/src/types/type.ts
export interface IEditFlashcard {
    _id: string
    title: string
    language: string
    desc: string
    public: boolean
}